pyproject.toml
[tool.poetry]
name = "acesso-livre-api"
version = "0.1.0"
description = ""
authors = ["Wandrey <wandrey7dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.116.2"}
pydantic-settings = {extras = ["dotenv"], version = "^2.10.1"}
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
python-jose = "^3.5.0"
passlib = "^1.7.4"
bcrypt = "<4.0"
alembic = "^1.17.1"

[tool.pytest.ini_options]
markers = ["integration: marca testes de integração"]

[tool.black]
line-length = 90
target-version = ['py311']

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-mock = "^3.15.1"
pytest-asyncio = "^1.2.0"
httpx = "^0.28.1"
pytest-cov = "^7.0.0"
pylint = "^4.0.2"
black = "^25.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
